+++ src/matterbridgeDevice.ts
@@ -4,7 +4,7 @@
  * @file matterbridgeDevice.ts
  * @author Luca Liguori
  * @date 2023-12-29
- * @version 1.1.1
+ * @version 1.2.0
  *
  * Copyright 2023, 2024 Luca Liguori.
  *
@@ -79,7 +79,7 @@ import { ClusterId, EndpointNumber, GroupId, VendorId } from '@project-chip/matt
 import { Device, DeviceClasses, DeviceTypeDefinition, Endpoint, EndpointOptions } from '@project-chip/matter-node.js/device';
 import { AtLeastOne, extendPublicHandlerMethods } from '@project-chip/matter-node.js/util';
 
-import { EveHistory, EveHistoryCluster, MatterHistory, Sensitivity, WeatherTrend, TemperatureDisplayUnits } from 'matter-history';
+import { EveHistory, MatterHistory } from 'matter-history';
 import { AnsiLogger, CYAN, LogLevel, TimestampFormat, YELLOW, db, debugStringify, hk, or, zb } from 'node-ansi-logger';
 
 import { AirQuality, AirQualityCluster } from './cluster/AirQualityCluster.js';
@@ -536,7 +536,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
     if (includeServerList.includes(Scenes.Cluster.id)) endpoint.addClusterServer(this.getDefaultScenesClusterServer());
     if (includeServerList.includes(OnOff.Cluster.id)) endpoint.addClusterServer(this.getDefaultOnOffClusterServer());
     if (includeServerList.includes(LevelControl.Cluster.id)) endpoint.addClusterServer(this.getDefaultLevelControlClusterServer());
-    if (includeServerList.includes(ColorControl.Cluster.id)) endpoint.addClusterServer(this.getDefaultColorControlClusterServer());
+    if (includeServerList.includes(ColorControl.Cluster.id)) endpoint.addClusterServer(this.getDefaultCompleteColorControlClusterServer());
     if (includeServerList.includes(Switch.Cluster.id)) endpoint.addClusterServer(this.getDefaultSwitchClusterServer());
     if (includeServerList.includes(DoorLock.Cluster.id)) endpoint.addClusterServer(this.getDefaultDoorLockClusterServer());
     if (includeServerList.includes(Thermostat.Cluster.id)) endpoint.addClusterServer(this.getDefaultThermostatClusterServer());
@@ -686,24 +686,24 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    * @param {Endpoint} [endpoint] - (Optional) The endpoint to set the attribute on. If not provided, the attribute will be set on the current endpoint.
    */
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  setAttribute(clusterId: ClusterId, attribute: string, value: any, log?: AnsiLogger, endpoint?: Endpoint) {
+  setAttribute(clusterId: ClusterId, attribute: string, value: any, log?: AnsiLogger, endpoint?: Endpoint): boolean {
     if (!endpoint) endpoint = this as Endpoint;
 
     const clusterServer = endpoint.getClusterServerById(clusterId);
     if (!clusterServer) {
       log?.error(`setAttribute error: Cluster ${clusterId} not found on endpoint ${endpoint.name}:${endpoint.number}`);
-      return;
+      return false;
     }
     const capitalizedAttributeName = attribute.charAt(0).toUpperCase() + attribute.slice(1);
     if (!clusterServer.isAttributeSupportedByName(attribute) && !clusterServer.isAttributeSupportedByName(capitalizedAttributeName)) {
       if (log) log.error(`setAttribute error: Attribute ${attribute} not found on Cluster ${clusterId} on endpoint ${endpoint.name}:${endpoint.number}`);
-      return;
+      return false;
     }
     // Find the getter method
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     if (!(clusterServer as any)[`get${capitalizedAttributeName}Attribute`]) {
       log?.error(`setAttribute error: Getter get${capitalizedAttributeName}Attribute not found on Cluster ${clusterServer.name} on endpoint ${endpoint.name}:${endpoint.number}`);
-      return;
+      return false;
     }
     // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-empty-object-type
     const getter = (clusterServer as any)[`get${capitalizedAttributeName}Attribute`] as () => {};
@@ -711,7 +711,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     if (!(clusterServer as any)[`set${capitalizedAttributeName}Attribute`]) {
       log?.error(`setAttribute error: Setter set${capitalizedAttributeName}Attribute not found on Cluster ${clusterServer.name} on endpoint ${endpoint.name}:${endpoint.number}`);
-      return;
+      return false;
     }
     // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-empty-object-type
     const setter = (clusterServer as any)[`set${capitalizedAttributeName}Attribute`] as (value: any) => {};
@@ -722,6 +722,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
         `from ${YELLOW}${typeof oldValue === 'object' ? debugStringify(oldValue) : oldValue}${db} ` +
         `to ${YELLOW}${typeof value === 'object' ? debugStringify(value) : value}${db}`,
     );
+    return true;
   }
 
   /**
@@ -797,6 +798,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    * @param consumption - The consumption value (default: 0).
    * @returns The default static EveHistoryClusterServer object.
    */
+  /*
   getDefaultStaticEveHistoryClusterServer(voltage = 0, current = 0, power = 0, consumption = 0) {
     return ClusterServer(
       EveHistoryCluster.with(EveHistory.Feature.EveEnergy),
@@ -823,6 +825,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
       {},
     );
   }
+  */
 
   /**
    * Create a default static EveHistoryClusterServer object with the specified voltage, current, power, and consumption values.
@@ -834,9 +837,11 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    * @param consumption - The consumption value (default: 0).
    * @returns The default static EveHistoryClusterServer object.
    */
+  /*
   createDefaultStaticEveHistoryClusterServer(voltage = 0, current = 0, power = 0, consumption = 0) {
     this.addClusterServer(this.getDefaultStaticEveHistoryClusterServer(voltage, current, power, consumption));
   }
+  */
 
   /**
    * Creates a room Eve History Cluster Server.
@@ -845,6 +850,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    * @param history - The MatterHistory object.
    * @param log - The AnsiLogger object.
    */
+  /*
   createRoomEveHistoryClusterServer(history: MatterHistory, log: AnsiLogger) {
     history.setMatterHystoryType('room', this.serialNumber);
     this.addClusterServer(
@@ -909,6 +915,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
       ),
     );
   }
+  */
 
   /**
    * Creates a Weather Eve History Cluster Server.
@@ -917,6 +924,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    * @param history - The MatterHistory instance.
    * @param log - The AnsiLogger instance.
    */
+  /*
   createWeatherEveHistoryClusterServer(history: MatterHistory, log: AnsiLogger) {
     history.setMatterHystoryType('weather', this.serialNumber);
     this.addClusterServer(
@@ -984,6 +992,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
       ),
     );
   }
+  */
 
   /**
    * Creates an Energy Eve History Cluster Server.
@@ -992,6 +1001,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    * @param history - The MatterHistory object.
    * @param log - The AnsiLogger object.
    */
+  /*
   createEnergyEveHistoryClusterServer(history: MatterHistory, log: AnsiLogger) {
     history.setMatterHystoryType('energy');
     this.addClusterServer(
@@ -1077,6 +1087,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
       ),
     );
   }
+  */
 
   /**
    * Creates a Motion Eve History Cluster Server.
@@ -1085,6 +1096,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    * @param history - The MatterHistory object.
    * @param log - The AnsiLogger object.
    */
+  /*
   createMotionEveHistoryClusterServer(history: MatterHistory, log: AnsiLogger) {
     history.setMatterHystoryType('motion');
     this.addClusterServer(
@@ -1155,6 +1167,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
       ),
     );
   }
+  */
 
   /**
    * Creates a door EveHistoryCluster server.
@@ -1163,6 +1176,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    * @param history - The MatterHistory instance.
    * @param log - The AnsiLogger instance.
    */
+  /*
   createDoorEveHistoryClusterServer(history: MatterHistory, log: AnsiLogger) {
     history.setMatterHystoryType('door');
     this.addClusterServer(
@@ -1248,6 +1262,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
       ),
     );
   }
+  */
 
   /**
    * Get a default IdentifyCluster server.
@@ -1773,6 +1788,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
 
   /**
    * Get a default color control cluster server.
+   * @deprecated This method is deprecated and will be removed in a future version. Use getDefaultCompleteColorControlClusterServer.
    *
    * @param currentHue - The current hue value.
    * @param currentSaturation - The current saturation value.
@@ -1846,6 +1862,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
   }
   /**
    * Creates a default color control cluster server.
+   * @deprecated This method is deprecated and will be removed in a future version. Use createDefaultCompleteColorControlClusterServer.
    *
    * @param currentHue - The current hue value.
    * @param currentSaturation - The current saturation value.
@@ -1859,6 +1876,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
 
   /**
    * Get a default color control cluster server.
+   * @deprecated This method is deprecated and will be removed in a future version. Use getDefaultCompleteColorControlClusterServer.
    *
    * @param currentX - The current X value.
    * @param currentY - The current Y value.
@@ -1941,6 +1959,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
   }
   /**
    * Creates a default color control cluster server.
+   * @deprecated This method is deprecated and will be removed in a future version. Use createDefaultCompleteColorControlClusterServer.
    *
    * @param currentX - The current X value.
    * @param currentY - The current Y value.
@@ -1952,6 +1971,103 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    */
   createDefaultXYColorControlClusterServer(currentX = 0, currentY = 0, colorTemperatureMireds = 500, colorTempPhysicalMinMireds = 147, colorTempPhysicalMaxMireds = 500) {
     this.addClusterServer(this.getDefaultXYColorControlClusterServer(currentX, currentY, colorTemperatureMireds, colorTempPhysicalMinMireds, colorTempPhysicalMaxMireds));
+  }
+
+  /**
+   * Get a default color control cluster server.
+   *
+   * @param currentX - The current X value.
+   * @param currentY - The current Y value.
+   * @param currentHue - The current hue value.
+   * @param currentSaturation - The current saturation value.
+   * @param colorTemperatureMireds - The color temperature in mireds.
+   * @param colorTempPhysicalMinMireds - The physical minimum color temperature in mireds.
+   * @param colorTempPhysicalMaxMireds - The physical maximum color temperature in mireds.
+   */
+  getDefaultCompleteColorControlClusterServer(currentX = 0, currentY = 0, colorTemperatureMireds = 500, colorTempPhysicalMinMireds = 147, colorTempPhysicalMaxMireds = 500) {
+    return ClusterServer(
+      ColorControlCluster.with(ColorControl.Feature.Xy, ColorControl.Feature.HueSaturation, ColorControl.Feature.ColorTemperature),
+      {
+        colorMode: ColorControl.ColorMode.CurrentHueAndCurrentSaturation,
+        options: {
+          executeIfOff: false,
+        },
+        numberOfPrimaries: null,
+        enhancedColorMode: ColorControl.EnhancedColorMode.CurrentHueAndCurrentSaturation,
+        colorCapabilities: { xy: true, hueSaturation: true, colorLoop: false, enhancedHue: false, colorTemperature: true },
+        currentHue: 0,
+        currentSaturation: 0,
+        currentX,
+        currentY,
+        colorTemperatureMireds,
+        colorTempPhysicalMinMireds,
+        colorTempPhysicalMaxMireds,
+      },
+      {
+        moveToColor: async (data) => {
+          this.log.debug('Matter command: moveToColor request:', data.request, 'attributes.currentHue:', data.attributes.currentX.getLocal(), data.attributes.currentY.getLocal());
+          this.commandHandler.executeHandler('moveToColor', data);
+        },
+        moveColor: async () => {
+          this.log.error('Matter command: moveColor not implemented');
+        },
+        stepColor: async () => {
+          this.log.error('Matter command: stepColor not implemented');
+        },
+        moveToHue: async ({ request, attributes, endpoint }) => {
+          this.log.debug('Matter command: moveToHue request:', request, 'attributes.currentHue:', attributes.currentHue.getLocal());
+          this.commandHandler.executeHandler('moveToHue', { request, attributes, endpoint });
+        },
+        moveHue: async () => {
+          this.log.error('Matter command: moveHue not implemented');
+        },
+        stepHue: async () => {
+          this.log.error('Matter command: stepHue not implemented');
+        },
+        moveToSaturation: async ({ request, attributes, endpoint }) => {
+          this.log.debug('Matter command: moveToSaturation request:', request, 'attributes.currentSaturation:', attributes.currentSaturation.getLocal());
+          this.commandHandler.executeHandler('moveToSaturation', { request, attributes, endpoint });
+        },
+        moveSaturation: async () => {
+          this.log.error('Matter command: moveSaturation not implemented');
+        },
+        stepSaturation: async () => {
+          this.log.error('Matter command: stepSaturation not implemented');
+        },
+        moveToHueAndSaturation: async ({ request, attributes, endpoint }) => {
+          this.log.debug('Matter command: moveToHueAndSaturation request:', request, 'attributes.currentHue:', attributes.currentHue.getLocal(), 'attributes.currentSaturation:', attributes.currentSaturation.getLocal());
+          this.commandHandler.executeHandler('moveToHueAndSaturation', { request, attributes, endpoint });
+        },
+        stopMoveStep: async () => {
+          this.log.error('Matter command: stopMoveStep not implemented');
+        },
+        moveToColorTemperature: async ({ request, attributes, endpoint }) => {
+          this.log.debug('Matter command: moveToColorTemperature request:', request, 'attributes.colorTemperatureMireds:', attributes.colorTemperatureMireds.getLocal());
+          this.commandHandler.executeHandler('moveToColorTemperature', { request, attributes, endpoint });
+        },
+        moveColorTemperature: async () => {
+          this.log.error('Matter command: moveColorTemperature not implemented');
+        },
+        stepColorTemperature: async () => {
+          this.log.error('Matter command: stepColorTemperature not implemented');
+        },
+      },
+      {},
+    );
+  }
+  /**
+   * Creates a default color control cluster server.
+   *
+   * @param currentX - The current X value.
+   * @param currentY - The current Y value.
+   * @param currentHue - The current hue value.
+   * @param currentSaturation - The current saturation value.
+   * @param colorTemperatureMireds - The color temperature in mireds.
+   * @param colorTempPhysicalMinMireds - The physical minimum color temperature in mireds.
+   * @param colorTempPhysicalMaxMireds - The physical maximum color temperature in mireds.
+   */
+  createDefaultCompleteColorControlClusterServer(currentX = 0, currentY = 0, colorTemperatureMireds = 500, colorTempPhysicalMinMireds = 147, colorTempPhysicalMaxMireds = 500) {
+    this.addClusterServer(this.getDefaultCompleteColorControlClusterServer(currentX, currentY, colorTemperatureMireds, colorTempPhysicalMinMireds, colorTempPhysicalMaxMireds));
   }
 
   /**
@@ -1994,7 +2110,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    */
   getDefaultWindowCoveringClusterServer(positionPercent100ths?: number) {
     return ClusterServer(
-      WindowCoveringCluster.with(WindowCovering.Feature.Lift, WindowCovering.Feature.PositionAwareLift, WindowCovering.Feature.AbsolutePosition),
+      WindowCoveringCluster.with(WindowCovering.Feature.Lift, WindowCovering.Feature.PositionAwareLift /* , WindowCovering.Feature.AbsolutePosition*/),
       {
         type: WindowCovering.WindowCoveringType.Rollershade,
         configStatus: {
@@ -2011,8 +2127,8 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
         mode: { motorDirectionReversed: false, calibrationMode: false, maintenanceMode: false, ledFeedback: false },
         targetPositionLiftPercent100ths: positionPercent100ths ?? 0, // 0 Fully open 10000 fully closed
         currentPositionLiftPercent100ths: positionPercent100ths ?? 0, // 0 Fully open 10000 fully closed
-        installedClosedLimitLift: 10000,
-        installedOpenLimitLift: 0,
+        // installedClosedLimitLift: 10000,
+        // installedOpenLimitLift: 0,
       },
       {
         upOrOpen: async (data) => {
@@ -2053,7 +2169,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    */
   setWindowCoveringTargetAsCurrentAndStopped(endpoint?: Endpoint) {
     if (!endpoint) endpoint = this as Endpoint;
-    const windowCoveringCluster = endpoint.getClusterServer(WindowCoveringCluster.with(WindowCovering.Feature.Lift, WindowCovering.Feature.PositionAwareLift, WindowCovering.Feature.AbsolutePosition));
+    const windowCoveringCluster = endpoint.getClusterServer(WindowCoveringCluster.with(WindowCovering.Feature.Lift, WindowCovering.Feature.PositionAwareLift));
     if (windowCoveringCluster) {
       const position = windowCoveringCluster.getCurrentPositionLiftPercent100thsAttribute();
       if (position !== null) {
@@ -2077,7 +2193,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    */
   setWindowCoveringCurrentTargetStatus(current: number, target: number, status: WindowCovering.MovementStatus, endpoint?: Endpoint) {
     if (!endpoint) endpoint = this as Endpoint;
-    const windowCoveringCluster = endpoint.getClusterServer(WindowCoveringCluster.with(WindowCovering.Feature.Lift, WindowCovering.Feature.PositionAwareLift, WindowCovering.Feature.AbsolutePosition));
+    const windowCoveringCluster = endpoint.getClusterServer(WindowCoveringCluster.with(WindowCovering.Feature.Lift, WindowCovering.Feature.PositionAwareLift));
     if (windowCoveringCluster) {
       windowCoveringCluster.setCurrentPositionLiftPercent100thsAttribute(current);
       windowCoveringCluster.setTargetPositionLiftPercent100thsAttribute(target);
@@ -2097,7 +2213,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    */
   setWindowCoveringStatus(status: WindowCovering.MovementStatus, endpoint?: Endpoint) {
     if (!endpoint) endpoint = this as Endpoint;
-    const windowCovering = endpoint.getClusterServer(WindowCoveringCluster.with(WindowCovering.Feature.Lift, WindowCovering.Feature.PositionAwareLift, WindowCovering.Feature.AbsolutePosition));
+    const windowCovering = endpoint.getClusterServer(WindowCoveringCluster.with(WindowCovering.Feature.Lift, WindowCovering.Feature.PositionAwareLift));
     if (!windowCovering) return;
     windowCovering.setOperationalStatusAttribute({ global: status, lift: status, tilt: status });
     this.log.debug(`Set WindowCovering operationalStatus: ${status}`);
@@ -2111,7 +2227,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    */
   getWindowCoveringStatus(endpoint?: Endpoint): WindowCovering.MovementStatus | undefined {
     if (!endpoint) endpoint = this as Endpoint;
-    const windowCovering = endpoint.getClusterServer(WindowCoveringCluster.with(WindowCovering.Feature.Lift, WindowCovering.Feature.PositionAwareLift, WindowCovering.Feature.AbsolutePosition));
+    const windowCovering = endpoint.getClusterServer(WindowCoveringCluster.with(WindowCovering.Feature.Lift, WindowCovering.Feature.PositionAwareLift));
     if (!windowCovering) return undefined;
     const status = windowCovering.getOperationalStatusAttribute();
     this.log.debug(`Get WindowCovering operationalStatus: ${status.global}`);
@@ -2255,14 +2371,18 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
    * @param {Endpoint} endpoint - The endpoint on which to trigger the event (default the device endpoint).
    * @returns {void}
    */
-  triggerSwitchEvent(event: 'Single' | 'Double' | 'Long' | 'Press' | 'Release', endpoint?: Endpoint, log?: AnsiLogger): void {
+  triggerSwitchEvent(event: 'Single' | 'Double' | 'Long' | 'Press' | 'Release', log?: AnsiLogger, endpoint?: Endpoint): boolean {
     if (!endpoint) endpoint = this as Endpoint;
 
     if (['Single', 'Double', 'Long'].includes(event)) {
       const cluster = endpoint.getClusterServer(SwitchCluster.with(Switch.Feature.MomentarySwitch, Switch.Feature.MomentarySwitchRelease, Switch.Feature.MomentarySwitchLongPress, Switch.Feature.MomentarySwitchMultiPress));
       if (!cluster || !cluster.getFeatureMapAttribute().momentarySwitch) {
         log?.error(`triggerSwitchEvent ${event} error: Switch cluster with MomentarySwitch not found on endpoint ${endpoint.name}:${endpoint.number}`);
-        return;
+        return false;
+      }
+      if (endpoint.number === undefined) {
+        log?.error(`triggerSwitchEvent ${event} error: Endpoint number not assigned on endpoint ${endpoint.name}:${endpoint.number}`);
+        return false;
       }
       if (event === 'Single') {
         cluster.setCurrentPositionAttribute(1);
@@ -2299,7 +2419,11 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
       const cluster = endpoint.getClusterServer(Switch.Complete);
       if (!cluster || !cluster.getFeatureMapAttribute().latchingSwitch) {
         log?.error(`triggerSwitchEvent ${event} error: Switch cluster with LatchingSwitch not found on endpoint ${endpoint.name}:${endpoint.number}`);
-        return;
+        return false;
+      }
+      if (endpoint.number === undefined) {
+        log?.error(`triggerSwitchEvent ${event} error: Endpoint number not assigned on endpoint ${endpoint.name}:${endpoint.number}`);
+        return false;
       }
       if (event === 'Press') {
         cluster.setCurrentPositionAttribute(1);
@@ -2314,6 +2438,7 @@ export class MatterbridgeDevice extends extendPublicHandlerMethods<typeof Device
         log?.info(`${db}Trigger endpoint ${or}${endpoint.name}:${endpoint.number}${db} event ${hk}${cluster.name}.Release${db}`);
       }
     }
+    return true;
   }
 
   /**