+++ CHANGELOG.md
@@ -4,8 +4,6 @@ All notable changes to this project will be documented in this file.
 
 If you like this project and find it useful, please consider giving it a star on GitHub at https://github.com/Luligu/matterbridge and sponsoring it.
 
-## [1.5.0] - 2024-08-27
-
 ### Breaking Changes
 
 - [-bridge -childbridge]: You don't need anymore to add the parmeter -bridge or -childbridge on the command line or systemctl configuration or docker command: the default is bridge mode and if no parameter is added, Matterbridge uses the settings from the frontend that are saved.
@@ -17,6 +15,22 @@ If you like this project and find it useful, please consider giving it a star on
 ### Breaking Changes for developers
 
 - please read this [Development guide lines](https://github.com/Luligu/matterbridge/blob/main/README-DEV.md)
+
+## [1.5.1] - 2024-08-28
+
+### Added
+
+### Changed
+
+- [matterbridgeDevice]: refactor WindowCovering cluster (removed AbsolutePosition).
+- [matterbridge]: Removed deprecated methods.
+- [package]: Update dependencies.
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="./yellow-button.png" alt="Buy me a coffee" width="120">
+</a>
+
+## [1.5.0] - 2024-08-27
 
 ### Added